+++ src/loaders/c_loader.c
@@ -11,7 +11,7 @@ typedef struct sk_c_mdata {
     void* handle;
 } sk_c_mdata;
 
-const char* sk_c_module_name(const char* short_name, char* fullname, int sz)
+const char* sk_c_module_name(const char* short_name, char* fullname, size_t sz)
 {
     memset(fullname, 0, sz);
     snprintf(fullname, sz, "modules/%s.so", short_name);
